Keep the settings pagination offset 64-bit
LegacyPagination.Params carries page/limit/offset as long, and every other
store in the port takes them as long. LegacySettingsStore was the lone int
signature, so the two catalogue controllers narrowed at the call.

At limit=100 a page above 21474837 produces an offset past Integer.MAX_VALUE.
The cast wraps it negative, MariaDB rejects the OFFSET outright, and the route
answers 500 where PHP -- computing in 64-bit throughout -- returns a successful
empty page.

+++ backend/src/main/java/com/workin/legacy/settings/LegacySettingsController.java
@@ -178,7 +178,7 @@ public class LegacySettingsController {
 		String locale = locale(request);
 		List<Map<String, Object>> rows = new ArrayList<>();
 		for (Map<String, Object> row : store.definitionRows(
-				search, (int) pagination.limit(), (int) pagination.offset())) {
+				search, pagination.limit(), pagination.offset())) {
 			Map<String, Object> shaped = new LinkedHashMap<>(row);
 			shaped.put("label", LegacySettingLabels.pick(locale,
 					nullableText(row.get("label_ar")), nullableText(row.get("label_en")),
@@ -218,7 +218,7 @@ public class LegacySettingsController {
 		String locale = locale(request);
 		List<Map<String, Object>> rows = new ArrayList<>();
 		for (Map<String, Object> row : store.allowedValueRows(
-				definitionId, (int) pagination.limit(), (int) pagination.offset())) {
+				definitionId, pagination.limit(), pagination.offset())) {
 			Map<String, Object> shaped = new LinkedHashMap<>(row);
 			shaped.put("label", LegacySettingLabels.pick(locale,
 					nullableText(row.get("label_ar")), nullableText(row.get("label_en")),
